"Added styling for home page, cart, login, and register views; updated server.js to include route handlers for new views"
This commit includes changes to both the front-end and back-end of the application. On the front-end, there are significant updates to the home page, as well as the addition of new views for the cart, login, and register pages. These updates are reflected in the changes to the HTML, CSS, and SCSS files. On the back-end, the server.js file has been updated to include route handlers for the new views. The commit message is concise and accurately describes the changes made in the code diff.